Split SetText into SetTextA/SetTextW to help understand string usage
While we have two string kinds mixed around (ascii and unicode) it is
better to use separate functions instead overloaded implementations.
This helps understand where which kind of string is used and their
flow in code so we can easily refactor this into kinds of strings for
display or packet and unify the string usage internaly over utf8.