Fixes network connection issue on Linux
The first login packet (0xef) sends a seed to the server in the
following 4 bytes, this seed is composed of the local ip address, but
the linux version was returning 0.0.0.0 and the server was rejecting it
by simple disconnecting. On the server logs this showed a message saying
'Invalid client detected, disconnecting'.