Initial string work for linux compatibility reading from packets
Unicode and Big-endian Unicode strings in packets are always 16bits, but
wchar_t is not compatible between windows and linux. This fixes some
cases where a direct buffer conversion to a wchar_t* were incorrectly
done.

This also removes a lot of redundant string allocations, there is still
a lot more string cleanup to be done, but right now this runs corectly
on linux and we're able to fully connect to a server.