Makes OrionUO client a shared object (DLL) on Linux only
This is needed because Orion Assistant (Qt) wants to own the main thread
of the application and loading Qt widgets in a dll does not work on other
platforms than Windows. To solve this and get Orion Assistant working on
Linux, the relation must be inversed, making OA the main application
that loads the client as a plugin. To maintain the client itself as an
application independent of OA, a client stub application that loads the
client plugin can be used. In practice this does not change much other
than things working or not working.