Deprecate BASS support even on Windows.
For the moment this means we do not have MIDI playback, this is being
worked on in a slower pace and will be fixed, but maybe not necessarely
before the porting effort finishes.
This also marks a point in the port progress, closing one of the biggest
issues with a true open source and free software implementation of the
Ultima client.