Try to split Data and Protocol Client Versions
This is confusing, and the use case is a bit weird and a special case.
Anything that may impact how the client iteract with the server uses Protocol
client version.
Anything related to loading data and interpreting files local in the user client
are Client version.
For normal use cases, this split should not have an impact as if Protocol
version is not set, then Client version will be used automatically.