Fixes new trade system issue with RunUO/ServUO.
The order and type of packets changed for the implementation of the
platinum currency system on these servers (not sure how it is on OSI).
This caused trade update packets to happen without a trade open and also
the creation of trade containers seem to have changed orders, causing
the trade to try to put items in the wrong containers and consequently
never drawing them.