Major cleanup
+ Fix CMake project on Windows that wasn't able to download
Dependencies;
+ Limit cotire to crossuo only;

Still a lot to be done, but this tries to reduce the monumental work
required to remove the precompiled header sickness.
Pro-tip: never ever ever use precompiled header in any project, it is
nuts!

Even if it is still not free of pch, here an overview of what was done:

- Cleanup of includes from hub headers (globals, pch, etc.), moved
closer to where they are actually needed.
- Forward declarations;
- Partial cleanup on globally using on std
  - string -> std::string
  - wstring -> std::wstring
  - etc.
- auto on string variables that can be auto to help future string
refactoring;
- Start removing the DEBUG_TRACE_FUNCTION macro;