Add support for server (incremental) mirroring
When using --mirror the output path will contain a certain structure to
keep track of versioning, eg:
  `/output/<product-timestamp>/<product-name>/cache/*`
a file `/output/latest.txt` will also be created to track the latest
version (product-timestamp) downloaded. If a new version is published,
this previous version will be used as a base for diff, and only changed
files will be downloaded in the new mirrored version.

The reason for doing this way is basically to enable full range of
different versions mirroring with less space posible when used with a
stackable filesystem like unionfs/uafs to serve files to an actual
UO.exe patcher.

So, xuoi acts as a server mirror tool and (in the future) an actual
client installer and patcher once the packaging is finished.

The mirroring functionality is helpful to reproduce different client
data versions as required for reversing / debugging needs.