CrossUO installer/updater reimplementation (first pass)
Implements the updater/installer module from X:UO Launcher as an
independent implementation to be reused/adapted as needed.

This is a first step towards an unified installer and tooling and
removing Qt as a dependency.

Some improvements to the fs.h utility:
- fs_copy file copying
- fs_del file deletion
- fs_chmod permissions manipulation
- fs_file_read utils
- fs_file_wrie utils

Separate http functions, add a helper function to download directly to
file instead of memory buffers