Move dependency management to vcpkg
This will unify all platforms and remove the need of a binary dependency
pack for windows. A side effect is that with vcpkg, every dependency is
statically built, this will help deploying on different linux distro and
easy install on osx/windows.

Removed cotire, causes too much issue with PCH.
Re-added the PCH header in some files to get it back building without
PCH, this will be cleared out after removing opengl includes.