Client: add more information on why client failed to launch
One of the biggest sources of technical support on the client are
users having difficulties launching the client for these reasons:
- No config was set
- No original client data found
- A config is found, but no valid client data is set/found

This tries to help differentiating the reasons above and try to help
users to fix the issues our a bit more detailed error message box.

If there is no config, the client will try to launch X:UO launcher and
hopefully the user will understand that a configuration is needed - more
help should be done in the launcher, as pointing to a website
documentation.
If then no launcher is found, the client will open the browser in the
crossuo website in the download section and ask the user to download the
launcher, which should be from now on always together with the CrossUO
client.
If the config is found, the client will try to validate that the UO data
path set in the config is correct and pop a message box explaining
and/or loading the X:UO Launcher.