Improve first time experience when building crossuo from source
Add a post-build step to copy xuolauncher to the same directory as the
crossuo client to improve the first time experience after building the
code from source. The reason is that once you build and try to launch
the client, there is no configuration file so crossuo will try to start
the launcher itself to create a new configuration. The client expects
the launcher to be on the same directory.

Signed-off-by: Danny Angelo Carminati Grein <fungos@gmail.com>